Fix for the addition of fractions
While not strictly necessary, having a get common denominator function might be useful, and this implementation works with fractions of the same denominator, so for now this will work. Later, because you shouldn't be allowed to add fractions of different denominators directly via the cauldron, we can add a check inside of the cauldron class in AddIngredient. That way, the ability to add fractions of different denominators is possible, just prevented because of how the rules of the game.